Rolling back changes to _get_dtype_input_and_vector

+++ ivy_tests/test_ivy/test_frontends/test_jax/test_numpy/test_mathematical_functions.py
@@ -113,7 +113,7 @@ def _get_castable_dtypes_values(draw, *, allow_nan=False, use_where=False):
 def _get_dtype_input_and_vector(draw):
     size1 = draw(helpers.ints(min_value=1, max_value=5))
     size2 = draw(helpers.ints(min_value=1, max_value=5))
-    dtype = draw(helpers.get_dtypes("numeric"))
+    dtype = draw(helpers.get_dtypes("integer"))
     vec1 = draw(helpers.array_values(dtype=dtype[0], shape=(size1, size2)))
     return dtype, vec1
